Adding new rules for R8 full mode support
