+++ buildSrc/src/main/kotlin/io/getstream/chat/android/Dependencies.kt
@@ -60,7 +60,7 @@ object Versions {
     internal const val MOCKITO_KOTLIN = "4.0.0"
     internal const val MOCKITO = "4.3.1"
     internal const val MOSHI = "1.13.0"
-    internal const val NAVIGATION = "2.3.5"
+    internal const val NAVIGATION = "2.4.2"
     internal const val OK2CURL = "0.7.0"
     internal const val OKHTTP = "4.9.3"
     internal const val PHOTOVIEW = "2.3.0"
