+++ gradle.properties
@@ -2,3 +2,4 @@ android.useAndroidX=true
 org.gradle.jvmargs=-Xmx2048m
 org.gradle.parallel=true
 org.gradle.caching=true
+android.enableR8.fullMode=true
