+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/endpoint/ConfigApi.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.getstream.chat.android.client.api2
+package io.getstream.chat.android.client.api2.endpoint
 
 import io.getstream.chat.android.client.api.AuthenticatedApi
 import io.getstream.chat.android.client.api2.model.response.AppSettingsResponse
