+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/endpoint/DeviceApi.kt
@@ -14,40 +14,37 @@
  * limitations under the License.
  */
 
-package io.getstream.chat.android.client.api2
+package io.getstream.chat.android.client.api2.endpoint
 
 import io.getstream.chat.android.client.api.AuthenticatedApi
 import io.getstream.chat.android.client.api.QueryParams
-import io.getstream.chat.android.client.api2.model.requests.PartialUpdateUsersRequest
-import io.getstream.chat.android.client.api2.model.requests.QueryUsersRequest
-import io.getstream.chat.android.client.api2.model.requests.UpdateUsersRequest
-import io.getstream.chat.android.client.api2.model.response.UpdateUsersResponse
-import io.getstream.chat.android.client.api2.model.response.UsersResponse
+import io.getstream.chat.android.client.api2.model.requests.AddDeviceRequest
+import io.getstream.chat.android.client.api2.model.response.CompletableResponse
+import io.getstream.chat.android.client.api2.model.response.DevicesResponse
 import io.getstream.chat.android.client.call.RetrofitCall
 import retrofit2.http.Body
+import retrofit2.http.DELETE
 import retrofit2.http.GET
-import retrofit2.http.PATCH
 import retrofit2.http.POST
 import retrofit2.http.Query
 
 @AuthenticatedApi
-internal interface UserApi {
-    @POST("/users")
-    fun updateUsers(
-        @Query(QueryParams.CONNECTION_ID) connectionId: String,
-        @Body body: UpdateUsersRequest,
-    ): RetrofitCall<UpdateUsersResponse>
+internal interface DeviceApi {
 
-    @PATCH("/users")
-    @JvmSuppressWildcards // See issue: https://github.com/square/retrofit/issues/3275
-    fun partialUpdateUsers(
+    @GET("/devices")
+    fun getDevices(
         @Query(QueryParams.CONNECTION_ID) connectionId: String,
-        @Body body: PartialUpdateUsersRequest,
-    ): RetrofitCall<UpdateUsersResponse>
+    ): RetrofitCall<DevicesResponse>
 
-    @GET("/users")
-    fun queryUsers(
+    @POST("devices")
+    fun addDevices(
         @Query(QueryParams.CONNECTION_ID) connectionId: String,
-        @UrlQueryPayload @Query("payload") payload: QueryUsersRequest,
-    ): RetrofitCall<UsersResponse>
+        @Body request: AddDeviceRequest,
+    ): RetrofitCall<CompletableResponse>
+
+    @DELETE("/devices")
+    fun deleteDevice(
+        @Query("id") deviceId: String,
+        @Query(QueryParams.CONNECTION_ID) connectionId: String,
+    ): RetrofitCall<CompletableResponse>
 }
